bitcoin_stockmem/requirements.txt
google-generativeai>=0.8.0
sentence-transformers>=3.0.0
torch>=2.0.0
transformers>=4.40.0
ccxt>=4.0.0
pycoingecko>=3.1.0
feedparser>=6.0.0
trafilatura>=1.6.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
